requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
sentence-transformers==3.2.1
transformers==4.46.3
accelerate==1.1.1
datasets==3.1.0
peft==0.13.2
trl==0.12.1
torch>=2.2.0
pydantic==2.9.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0

streamlit==1.41.1
